requirements.txt
emoji
pyrofork
tgcrypto
async_pymongo
pymongo
python-dotenv
requests
beautifulsoup4
aiohttp
chevron
cv-3
gTTS
regex
apscheduler==3.10.4
pytz
pykeyboard
pySmartDL
psutil
python-dateutil
telegraph
hachoir
Pillow
httpx[http2]
git+https://github.com/yasirarism/vcsi
deep-translator
telethon
pyrate_limiter
cachetools
cloudscraper
GitPython
aiofiles
uvloop
lxml_html_clean
bing_image_downloader
librosa
numpy
mutagen
matplotlib
uvicorn
lxml_html_clean
python-multipart
standard-imghdr